Refactor user service interfaces to user subpackage
Moved IUserService, IGetUserService, IRegisterUserService, IUpdateUserService, and IContextService interfaces and their implementations to the new 'services.user' subpackage. Updated all relevant imports and documentation references to reflect the new package structure for improved code organization.
