+++ docs/java_doc/apidocs/com/platform/user_service/services/class-use/IGetUserService.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (17) on Sun Oct 05 21:33:17 ART 2025 -->
-<title>Uses of Interface com.platform.user_service.services.IGetUserService (user-service 0.0.1-SNAPSHOT API)</title>
+<title>Uses of Interface com.platform.user_service.services.user.IGetUserService (user-service 0.0.1-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta name="dc.created" content="2025-10-05">
@@ -49,7 +49,7 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 title="Uses of Interface com.platform.user_service.services.IGetUserService" class="title">Uses of Interface<br>com.platform.user_service.services.IGetUserService</h1>
+<h1 title="Uses of Interface com.platform.user_service.services.user.IGetUserService" class="title">Uses of Interface<br>com.platform.user_service.services.user.IGetUserService</h1>
 </div>
 <div class="caption"><span>Packages that use <a href="../IGetUserService.html" title="interface in com.platform.user_service.services">IGetUserService</a></span></div>
 <div class="summary-table two-column-summary">
