+++ docs/java_doc/apidocs/com/platform/user_service/services/impl/ContextService.html
@@ -73,7 +73,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class ContextService" class="title">Class ContextService</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">com.platform.user_service.services.impl.ContextService</div>
+<div class="inheritance">com.platform.user_service.services.user.impl.ContextService</div>
 </div>
 <section class="class-description" id="class-description">
 <dl class="notes">
