+++ docs/java_doc/apidocs/com/platform/user_service/services/impl/GetUserService.html
@@ -73,7 +73,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class GetUserService" class="title">Class GetUserService</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">com.platform.user_service.services.impl.GetUserService</div>
+<div class="inheritance">com.platform.user_service.services.user.impl.GetUserService</div>
 </div>
 <section class="class-description" id="class-description">
 <dl class="notes">
