+++ docs/java_doc/apidocs/com/platform/user_service/services/impl/class-use/ContextService.html
@@ -49,7 +49,7 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 title="Uses of Class com.platform.user_service.services.impl.ContextService" class="title">Uses of Class<br>com.platform.user_service.services.impl.ContextService</h1>
+<h1 title="Uses of Class com.platform.user_service.services.user.impl.ContextService" class="title">Uses of Class<br>com.platform.user_service.services.impl.ContextService</h1>
 </div>
 No usage of com.platform.user_service.services.impl.ContextService</main>
 <footer role="contentinfo">
