+++ docs/java_doc/apidocs/com/platform/user_service/services/impl/class-use/RegisterUserService.html
@@ -49,7 +49,7 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 title="Uses of Class com.platform.user_service.services.impl.RegisterUserService" class="title">Uses of Class<br>com.platform.user_service.services.impl.RegisterUserService</h1>
+<h1 title="Uses of Class com.platform.user_service.services.user.impl.RegisterUserService" class="title">Uses of Class<br>com.platform.user_service.services.impl.RegisterUserService</h1>
 </div>
 No usage of com.platform.user_service.services.impl.RegisterUserService</main>
 <footer role="contentinfo">
