+++ docs/java_doc/apidocs/com/platform/user_service/services/impl/class-use/UserService.html
@@ -49,7 +49,7 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 title="Uses of Class com.platform.user_service.services.impl.UserService" class="title">Uses of Class<br>com.platform.user_service.services.impl.UserService</h1>
+<h1 title="Uses of Class com.platform.user_service.services.user.impl.UserService" class="title">Uses of Class<br>com.platform.user_service.services.impl.UserService</h1>
 </div>
 No usage of com.platform.user_service.services.impl.UserService</main>
 <footer role="contentinfo">
