+++ src/main/java/com/platform/user_service/services/Ngo/Impl/NgoRegisterService.java
@@ -11,8 +11,8 @@ import com.platform.user_service.enums.NgoDocumentStatus;
 import com.platform.user_service.enums.NgoStatus;
 import com.platform.user_service.enums.UserRole;
 import com.platform.user_service.repositories.NgoRepository;
-import com.platform.user_service.services.IContextService;
-import com.platform.user_service.services.IUpdateUserService;
+import com.platform.user_service.services.user.IContextService;
+import com.platform.user_service.services.user.IUpdateUserService;
 import com.platform.user_service.services.Ngo.INgoRegisterService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
