+++ src/main/java/com/platform/user_service/services/Ngo/Impl/NgoUpdateService.java
@@ -9,7 +9,7 @@ import com.platform.user_service.entities.NgoImageEntity;
 import com.platform.user_service.enums.NgoDocumentStatus;
 import com.platform.user_service.enums.NgoStatus;
 import com.platform.user_service.repositories.NgoRepository;
-import com.platform.user_service.services.IContextService;
+import com.platform.user_service.services.user.IContextService;
 import com.platform.user_service.services.Ngo.INgoUpdateService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
