+++ src/main/java/com/platform/user_service/services/user/IContextService.java
@@ -1,4 +1,4 @@
-package com.platform.user_service.services;
+package com.platform.user_service.services.user;
 
 import org.springframework.stereotype.Service;
