+++ src/main/java/com/platform/user_service/services/user/impl/GetUserService.java
@@ -1,12 +1,12 @@
-package com.platform.user_service.services.impl;
+package com.platform.user_service.services.user.impl;
 
 import com.platform.user_service.controllers.manageExceptions.CustomException;
 import com.platform.user_service.dtos.common.UserDto;
 import com.platform.user_service.dtos.response.UserLoginResponseDto;
 import com.platform.user_service.entities.UserEntity;
 import com.platform.user_service.repositories.UserRepository;
-import com.platform.user_service.services.IContextService;
-import com.platform.user_service.services.IGetUserService;
+import com.platform.user_service.services.user.IContextService;
+import com.platform.user_service.services.user.IGetUserService;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
