+++ src/main/java/com/platform/user_service/services/user/impl/UpdateUserService.java
@@ -1,4 +1,4 @@
-package com.platform.user_service.services.impl;
+package com.platform.user_service.services.user.impl;
 
 import com.platform.user_service.controllers.manageExceptions.CustomException;
 import com.platform.user_service.dtos.request.UserUpdateDto;
@@ -7,8 +7,8 @@ import com.platform.user_service.entities.UserRoleEntity;
 import com.platform.user_service.entities.embeddable.UserRolId;
 import com.platform.user_service.enums.UserRole;
 import com.platform.user_service.repositories.UserRepository;
-import com.platform.user_service.services.IContextService;
-import com.platform.user_service.services.IUpdateUserService;
+import com.platform.user_service.services.user.IContextService;
+import com.platform.user_service.services.user.IUpdateUserService;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
